+++ src/pages/landing/landing-page.tsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router"
 
 import { FirstPointImage, SecondPointImage } from "@/shared/assets/images"
-import Button from "@/shared/design-system/button/button"
+import { Button } from "@/shared/design-system/button"
 import { Header } from "@/widgets"
 
 export default function LandingPage() {
