+++ src/shared/design-system/button/index.ts
@@ -1,0 +1,1 @@
+export { default as Button } from "./button"
